WIP, fixing all the assertions
