+++ x-pack/agent/pkg/agent/program/program_test.go
@@ -1,10 +1,17 @@
 package program
 
 import (
+	"io/ioutil"
+	"path/filepath"
+	"strings"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	yaml "gopkg.in/yaml.v2"
 
+	"github.com/elastic/beats/x-pack/agent/pkg/agent/internal/yamltest"
 	"github.com/elastic/beats/x-pack/agent/pkg/agent/transpiler"
 )
 
@@ -453,4 +460,95 @@ func TestGroupBy(t *testing.T) {
 		require.True(t, ok)
 		require.Equal(t, c2.Hash(), infosec1Config.Hash())
 	})
+}
+
+func TestConfiguration(t *testing.T) {
+	testcases := map[string]struct {
+		programs []string
+		expected int
+		err      bool
+	}{
+		"single_config": {
+			programs: []string{"filebeat", "metricbeat"},
+			expected: 2,
+		},
+		"audit_config": {
+			programs: []string{"auditbeat"},
+			expected: 1,
+		},
+		"journal_config": {
+			programs: []string{"journalbeat"},
+			expected: 1,
+		},
+		"monitor_config": {
+			programs: []string{"heartbeat"},
+			expected: 1,
+		},
+		"enabled_true": {
+			programs: []string{"filebeat"},
+			expected: 1,
+		},
+		"enabled_false": {
+			expected: 0,
+		},
+		"enabled_output_true": {
+			programs: []string{"filebeat"},
+			expected: 1,
+		},
+		"enabled_output_false": {
+			expected: 0,
+		},
+		"multiple_output_true": {
+			err: true,
+		},
+	}
+
+	for name, test := range testcases {
+		t.Run(name, func(t *testing.T) {
+			singleConfig, err := ioutil.ReadFile(filepath.Join("testdata", name+".yml"))
+			require.NoError(t, err)
+
+			var m map[string]interface{}
+			err = yaml.Unmarshal(singleConfig, &m)
+			require.NoError(t, err)
+
+			ast, err := transpiler.NewAST(m)
+			require.NoError(t, err)
+
+			programs, err := Programs(ast)
+			if test.err {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+
+			require.Equal(t, 1, len(programs))
+
+			defPrograms, ok := programs["default"]
+			require.True(t, ok)
+			require.Equal(t, test.expected, len(defPrograms))
+
+			for _, program := range defPrograms {
+				programConfig, err := ioutil.ReadFile(filepath.Join(
+					"testdata",
+					name+"-"+strings.ToLower(program.Spec.Name)+".yml",
+				))
+
+				require.NoError(t, err)
+				var m map[string]interface{}
+				err = yamltest.FromYAML(programConfig, &m)
+				require.NoError(t, err)
+
+				compareMap := &transpiler.MapVisitor{}
+				program.Config.Accept(compareMap)
+
+				if !assert.True(t, cmp.Equal(m, compareMap.Content)) {
+					diff := cmp.Diff(m, compareMap.Content)
+					if diff != "" {
+						t.Errorf("%s-%s mismatch (-want +got):\n%s", name, program.Spec.Name, diff)
+					}
+				}
+			}
+		})
+	}
 }
