+++ x-pack/agent/pkg/agent/program/spec_test.go
@@ -6,105 +6,16 @@ package program
 
 import (
 	"io/ioutil"
-	"path/filepath"
 	"regexp"
 	"strings"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 
-	"github.com/elastic/beats/x-pack/agent/pkg/agent/internal/yamltest"
 	"github.com/elastic/beats/x-pack/agent/pkg/agent/transpiler"
 )
-
-func TestConfiguration(t *testing.T) {
-	testcases := map[string]struct {
-		programs []string
-		expected int
-		err      bool
-	}{
-		"single_config": {
-			programs: []string{"filebeat", "metricbeat"},
-			expected: 2,
-		},
-		"audit_config": {
-			programs: []string{"auditbeat"},
-			expected: 1,
-		},
-		"journal_config": {
-			programs: []string{"journalbeat"},
-			expected: 1,
-		},
-		"monitor_config": {
-			programs: []string{"heartbeat"},
-			expected: 1,
-		},
-		"enabled_true": {
-			programs: []string{"filebeat"},
-			expected: 1,
-		},
-		"enabled_false": {
-			expected: 0,
-		},
-		"enabled_output_true": {
-			programs: []string{"filebeat"},
-			expected: 1,
-		},
-		"enabled_output_false": {
-			expected: 0,
-		},
-		"multiple_output_true": {
-			err: true,
-		},
-	}
-
-	for name, test := range testcases {
-		t.Run(name, func(t *testing.T) {
-			singleConfig, err := ioutil.ReadFile(filepath.Join("testdata", name+".yml"))
-			require.NoError(t, err)
-
-			var m map[string]interface{}
-			err = yaml.Unmarshal(singleConfig, &m)
-			require.NoError(t, err)
-
-			ast, err := transpiler.NewAST(m)
-			require.NoError(t, err)
-
-			programs, err := Programs(ast)
-			if test.err {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.Equal(t, test.expected, len(programs))
-
-			for _, program := range programs {
-				programConfig, err := ioutil.ReadFile(filepath.Join(
-					"testdata",
-					name+"-"+strings.ToLower(program.Spec.Name)+".yml",
-				))
-
-				require.NoError(t, err)
-				var m map[string]interface{}
-				err = yamltest.FromYAML(programConfig, &m)
-				require.NoError(t, err)
-
-				compareMap := &transpiler.MapVisitor{}
-				program.Config.Accept(compareMap)
-
-				if !assert.True(t, cmp.Equal(m, compareMap.Content)) {
-					diff := cmp.Diff(m, compareMap.Content)
-					if diff != "" {
-						t.Errorf("%s-%s mismatch (-want +got):\n%s", name, program.Spec.Name, diff)
-					}
-				}
-			}
-		})
-	}
-}
 
 func TestSerialization(t *testing.T) {
 	spec := Spec{
