+++ x-pack/agent/pkg/agent/program/testdata/audit_config.yml
@@ -1,4 +1,11 @@
-inputs:
+outputs:
+  default:
+    type: "elasticsearch"
+  elasticsearch:
+    hosts: [127.0.0.1:9200, 127.0.0.1:9300]
+    username: elastic
+    password: changeme
+streams:
   - type: log/file
     ignore_older: 123s
   - type: audit/auditd
@@ -20,8 +27,3 @@ management:
   host: "localhost"
 config:
   reload: 123
-output:
-  elasticsearch:
-    hosts: [127.0.0.1:9200, 127.0.0.1:9300]
-    username: elastic
-    password: changeme
