+++ x-pack/agent/pkg/agent/program/testdata/enabled_output_false.yml
@@ -1,4 +1,4 @@
-inputs:
+streams:
   - type: event/file
     paths:
       - /var/log/hello1.log
@@ -7,8 +7,9 @@ management:
   host: "localhost"
 config:
   reload: 123
-output:
-  elasticsearch:
+outputs:
+  default:
+    type: elasticsearch
     enabled: false
     hosts: [127.0.0.1:9200, 127.0.0.1:9300]
     username: elastic
