+++ x-pack/agent/pkg/agent/program/testdata/multiple_output_true.yml
@@ -1,4 +1,4 @@
-inputs:
+streams:
   - type: event/file
     paths:
       - /var/log/hello1.log
@@ -7,12 +7,14 @@ management:
   host: "localhost"
 config:
   reload: 123
-output:
-  elasticsearch:
+outputs:
+  default:
+    type: elasticsearch
     enabled: true
     hosts: [127.0.0.1:9200, 127.0.0.1:9300]
     username: elastic
     password: changeme
-  logstash:
+  other:
+    type: logstash
     enabled: true
     hosts: [127.0.0.1:9200, 127.0.0.1:9300]
