+++ x-pack/agent/pkg/agent/program/testdata/single_config.yml
@@ -1,4 +1,4 @@
-inputs:
+streams:
   - type: metric/docker
     setting: one
   - type: metric/apache
@@ -11,8 +11,9 @@ management:
   host: "localhost"
 config:
   reload: 123
-output:
-  elasticsearch:
+outputs:
+  default:
+    type: elasticsearch
     hosts: [127.0.0.1:9200, 127.0.0.1:9300]
     username: elastic
     password: changeme
